+++ src/Component/ContentElement/GridStartElement.php
@@ -6,8 +6,8 @@
  * @package    contao-bootstrap
  * @subpackage Grid
  * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2017 netzmacht David Molineus. All rights reserved.
- * @license    https://github.com/contao-bootstrap/grid/blob/master/LICENSE LGPL 3.0
+ * @copyright  2017-2019 netzmacht David Molineus. All rights reserved.
+ * @license    https://github.com/contao-bootstrap/grid/blob/master/LICENSE LGPL 3.0-or-later
  * @filesource
  */
 
@@ -69,6 +69,7 @@ final class GridStartElement extends AbstractGridElement
         try {
             $provider = $this->getGridProvider();
             $iterator = $provider->getIterator('ce:' . $this->get('id'), (int) $this->get('bs_grid'));
+            $this->responseTagger->addTags(['contao.db.tl_bs_grid.' . $this->get('bs_grid')]);
 
             return $iterator;
         } catch (\Exception $e) {