+++ src/Component/ContentElement/GridStopElement.php
@@ -6,8 +6,8 @@
  * @package    contao-bootstrap
  * @subpackage Grid
  * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2017 netzmacht David Molineus. All rights reserved.
- * @license    https://github.com/contao-bootstrap/grid/blob/master/LICENSE LGPL 3.0
+ * @copyright  2017-2019 netzmacht David Molineus. All rights reserved.
+ * @license    https://github.com/contao-bootstrap/grid/blob/master/LICENSE LGPL 3.0-or-later
  * @filesource
  */
 
@@ -76,7 +76,10 @@ final class GridStopElement extends AbstractGridElement
 
         if ($parent) {
             try {
-                return $provider->getIterator('ce:' . $parent->id, (int) $parent->bs_grid);
+                $iterator = $provider->getIterator('ce:' . $parent->id, (int) $parent->bs_grid);
+                $this->responseTagger->addTags(['contao.db.tl_bs_grid.' . $parent->bs_grid]);
+
+                return $iterator;
             } catch (\Exception $e) {
                 // Do nothing. In backend view an error is shown anyway.
                 return null;