+++ src/Component/FormField/AbstractFormField.php
@@ -6,8 +6,8 @@
  * @package    contao-bootstrap
  * @subpackage Grid
  * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2017 netzmacht David Molineus. All rights reserved.
- * @license    https://github.com/contao-bootstrap/grid/blob/master/LICENSE LGPL 3.0
+ * @copyright  2017-2019 netzmacht David Molineus. All rights reserved.
+ * @license    https://github.com/contao-bootstrap/grid/blob/master/LICENSE LGPL 3.0-or-later
  * @filesource
  */
 
@@ -17,6 +17,7 @@ namespace ContaoBootstrap\Grid\Component\FormField;
 
 use Contao\Widget;
 use ContaoBootstrap\Grid\Component\ComponentTrait;
+use Netzmacht\Contao\Toolkit\Response\ResponseTagger;
 
 /**
  * Class AbstractGridElement.
@@ -26,4 +27,14 @@ use ContaoBootstrap\Grid\Component\ComponentTrait;
 abstract class AbstractFormField extends Widget
 {
     use ComponentTrait;
+
+    /**
+     * Get the response tagger service.
+     *
+     * @return ResponseTagger
+     */
+    protected function getResponseTagger(): ResponseTagger
+    {
+        return self::getContainer()->get('contao_bootstrap.grid.response_tagger');
+    }
 }